added fullness check to dosensorreadandstore()

+++ lib/I2CFram/I2CFram.h
@@ -6,8 +6,8 @@
 
 
 
-/// @brief Write 16 bytes of globalcache to data memory
-uint16_t framwritesensordata( uint8_t *globalcache );
+/// @brief Write 16 bytes of globalcache to data memory, return true if full
+uint8_t framwritesensordata( uint8_t *globalcache );
